chore: Move search API and remove duped tests
Close #2967
