+++ src/core/api/index.js
@@ -10,8 +10,8 @@ import { oneLine } from 'common-tags';
 import config from 'config';
 
 import { initialApiState } from 'core/reducers/api';
-import { ADDON_TYPE_THEME, CLIENT_APP_ANDROID } from 'core/constants';
 import log from 'core/logger';
+import searchApi from 'core/api/search';
 import { convertFiltersToQueryParams } from 'core/searchUtils';
 import type { ErrorHandlerType } from 'core/errorHandler';
 import type { ApiStateType } from 'core/reducers/api';
@@ -162,56 +162,6 @@ export function callApi({
     .then((response) => (schema ? normalize(response, schema) : response));
 }
 
-type SearchParams = {|
-  api: ApiStateType,
-  auth: boolean,
-  // TODO: Make a "searchFilters" type because these are the same args
-  // for convertFiltersToQueryParams.
-  filters: {|
-    addonType?: string,
-    clientApp?: string,
-    category?: string,
-    page?: number,
-    page_size?: number,
-    query?: string,
-    sort?: string,
-  |},
-|};
-
-export function search(
-  { api, auth = false, filters = {} }: SearchParams
-) {
-  const _filters = { ...filters };
-  if (!_filters.clientApp && api.clientApp) {
-    log.debug(
-      `No clientApp found in filters; using api.clientApp (${api.clientApp})`);
-    _filters.clientApp = api.clientApp;
-  }
-  // TODO: This loads Firefox personas (lightweight themes) for Android
-  // until github.com/mozilla/addons-frontend/issues/1723#issuecomment-278793546
-  // and https://github.com/mozilla/addons-server/issues/4766 are addressed.
-  // Essentially: right now there are no categories for the combo
-  // of "Android" + "Themes" but Firefox lightweight themes will work fine
-  // on mobile so we request "Firefox" + "Themes" for Android instead.
-  // Obviously we need to fix this on the API end so our requests aren't
-  // overridden, but for now this will work.
-  if (
-    _filters.clientApp === CLIENT_APP_ANDROID &&
-    _filters.addonType === ADDON_TYPE_THEME
-  ) {
-    log.info(oneLine`addonType: ${_filters.addonType}/clientApp:
-      ${_filters.clientApp} is not supported. Changing clientApp to "firefox"`);
-    _filters.clientApp = 'firefox';
-  }
-  return callApi({
-    endpoint: 'addons/search',
-    schema: { results: [addon] },
-    params: convertFiltersToQueryParams(_filters),
-    state: api,
-    auth,
-  });
-}
-
 type FetchAddonParams = {|
   api: ApiStateType,
   slug: string,
@@ -309,21 +259,4 @@ export function logOutFromServer({ api }: {| api: ApiStateType |}) {
   });
 }
 
-type AutocompleteParams = {|
-  api: ApiStateType,
-  filters: {|
-    query: string,
-    addonType?: string,
-  |},
-|};
-
-export function autocomplete({ api, filters }: AutocompleteParams) {
-  return callApi({
-    endpoint: 'addons/autocomplete',
-    params: {
-      app: api.clientApp,
-      ...convertFiltersToQueryParams(filters),
-    },
-    state: api,
-  });
-}
+export const search = searchApi;
